// src/generated/MemberView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MembershipView } from "./MembershipView";

export type MemberView = { user_id: string, display_name: string | null, avatar_url: string | null, power_level: number, membership: MembershipView, member_ts: number | null, kicked: boolean, };